+++ History.md
@@ -1,5 +1,13 @@
+1.1.4 / 2014-06-25
+==================
+
+ * fix: fix some bugs
+ * add: perfmjs.utils#fastBind, perfmjs.utils#forEach
+ * modify: joquery#where, joquery#updateOrInsert, joquery#last, joquery#last
+
+
 1.1.3 / 2014-06-20
 ==================
 
- * index: fix some bugs
- * add: perfmjs.utils中增加toNumber, toBoolean方法
+ * fix: fix some bugs
+ * add: perfmjs.utils#toNumber, perfmjs.utils#toBoolean